refactor: move route to datafields folder
